Full CRUD and forms for remaining n-tities

+++ ex-2/app/models/cart/Cart.scala
@@ -2,7 +2,7 @@ package models.cart
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Cart(id: Long, value: Double)
+case class Cart(id: Long, value: Int)
 
 object Cart {
   implicit val cartFormat: OFormat[Cart] = Json.format[Cart]
